% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata_2.R, R/add_metadata_3.R
\name{add_metadata}
\alias{add_metadata}
\alias{add_metadata}
\title{Add metadata to data.frame}
\usage{
add_metadata(x, metad, xid = NULL, mid = NULL, drop_mid = T)

add_metadata(x, metad, xid = NULL, mid = NULL, drop_mid = T)
}
\arguments{
\item{x}{Data frame with data of interest}

\item{metad}{Data frame with metadata that should be added; or phyloseq object with sample_data slot}

\item{xid}{Character, name of the column with row identifiers in the main data (x)}

\item{mid}{Character, name of the column with row identifiers in the metadata (metad)}

\item{drop_mid}{Logical; if TRUE (default), metadata identifiers will be removed from results}

\item{x}{Data frame with data of interest}

\item{metad}{Data frame with metadata that should be added; or phyloseq object with sample_data slot}

\item{xid}{Character, name of the column with row identifiers in the main data (x)}

\item{mid}{Character, name of the column with row identifiers in the metadata (metad)}

\item{drop_mid}{Logical; if TRUE (default), metadata identifiers will be removed from results}
}
\value{
Data frame

Data frame
}
\description{
This function combines two tables based on the row identifiers.

This function combines two tables based on the row identifiers.
}
\examples{


}
