% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis_pairwise.R
\name{adonis_pairwise}
\alias{adonis_pairwise}
\title{Pairwise comparisons for permutational multivariate analysis of variance using distance matrices}
\usage{
adonis_pairwise(x, dd, group.var = "Fact", permut = 999, p.adj = T,
  adj.meth = "fdr", all_results = T, comparison_sep = ".", ...)
}
\arguments{
\item{x}{Sample meta-data (data frame for the independent variables)}

\item{dd}{Dissimilarity matrix between samples}

\item{group.var}{Name of the independent variable to test (RHS in adonis formula)}

\item{permut}{Number of permutations required}

\item{p.adj}{Logical, adjust P-values for multiple comparisons}

\item{adj.meth}{Correction method from \code{\link{p.adjust}} ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr")}

\item{all_results}{Logical, return results of adonis and data subsets for each pairwise comparison}

\item{comparison_sep}{Character string to separate the levels of independent variable the in the pairwise comparison names (default, ".")}

\item{...}{Additional arguments will be passed to \code{\link{adonis}}}
}
\value{
List with ......
}
\examples{
library(vegan)
data(dune)
data(dune.env)

# Compare all Management levels
adonis(dune ~ Management, data = dune.env)

# Pairwise comparisons between Management levels
ad <- adonis_pairwise(x = dune.env, dd = vegdist(dune), group.var = "Management")
ad$Adonis.tab

}
\seealso{
\code{\link{adonis}}
}
