% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_to_wide.R
\name{blast_to_wide}
\alias{blast_to_wide}
\title{Convert BLAST results to a wide table}
\usage{
blast_to_wide(
  blst,
  max_hits = 10,
  taxonomy = NULL,
  seqs = NULL,
  refs = NULL,
  add_coverage = TRUE,
  remove_missing = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{blst}{Data.table with blast hits (in long format, from m8 file)}

\item{max_hits}{Number of targets to preserve in a table}

\item{taxonomy}{\itemize{
  \item{"NA"}, no taxonomy should be added to the table
  \item{"NULL"}, use default tax ranks (Kingdom - Species)
  \item{"custom character vector"} with column names that contain taxonomy information
}}

\item{seqs}{Query sequences (XStringSet object from Biostrings package)}

\item{refs}{Target sequences (XStringSet object from Biostrings package)}

\item{add_coverage}{Logical, add query coverage to the output (default, TRUE)}

\item{remove_missing}{Logical, remove blast matches if sequences are missing in `seqs` and `refs` (default, FALSE)}

\item{verbose}{Logical, produce more screen output (default, TRUE)}
}
\value{
data.table
}
\description{
Convert BLAST results to a wide table
}
\details{
For each query, matches should be sorted in the desired order (e.g., by BitScore or E-Value).
}
