% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Bubble plot}
\usage{
bubble_plot(x, transp = 0.9, circ = 16, add.dendro = FALSE, ...)
}
\arguments{
\item{x}{Data frame with data (columns = samples, rows = species abundance, rownames = species names)}

\item{transp}{Transparency level (0-1)}

\item{circ}{Circle scale factor}

\item{add.dendro}{Logical, adds dendrogram for samples}

\item{...}{}
}
\value{
Invisibly returns a plot ('ggplot' class).
}
\examples{
x <- as.data.frame(matrix(runif(n = 100, min = 0, max = 100), nrow = 10))
rownames(x) <- paste("sp", 1:nrow(x), sep="")
bubble_plot(x, add.dendro=F)
bubble_plot(x, add.dendro=T)

}
