% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Split a vector into chunks equal size}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{A vector (e.g., character or numeric vector)}

\item{n}{The number of chunks required (integer)}
}
\value{
List (of the length defined by the number of chunks) with vectors
}
\description{
Split a vector into chunks equal size
}
\examples{
chunk(x = letters, n = 4)

}
