% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_learn_errors_subset.R
\name{dada_learn_errors_subset}
\alias{dada_learn_errors_subset}
\title{DADA2 learning the error rates based on a subset of data}
\usage{
dada_learn_errors_subset(filts, n = 25, samps = NULL, multitr = TRUE,
  verbose = TRUE)
}
\arguments{
\item{filts}{character vector with the names of pre-filtered fastq or fastq.gz files}

\item{n}{integer, how many samples to choose for learning (will be ignored if samps != NULL)}

\item{samps}{character vector, which samples to take for learning (overrides n-argument)}

\item{multitr}{logical, enables multithreading}

\item{verbose}{logical, shows some additional information}
}
\value{
Matrix with estimated error rates that can be used with dada-function.
}
\description{
Function to learn the error rates from a subset of the data (for the big data workflow)
}
\examples{
library(dada2)
# filtF and filtR are character vectors with the full path to fastq.gz-files
errsF <- learn_errors_subset(filtF, n = 30)
errsR <- learn_errors_subset(filtR, samps = attr(errsF, "selected_samples"))
}
\references{
\url{https://benjjneb.github.io/dada2/bigdata.html}
}
\seealso{
\code{\link{dada}}
}
