% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfRowName.R
\name{dfRowName}
\alias{dfRowName}
\title{Transfer rownames of data frame to a new column (for data export)}
\usage{
dfRowName(x, name = "Rows")
}
\arguments{
\item{x}{Data frame}

\item{name}{Name of the new column to be created}
}
\value{
Data frame
}
\examples{
x <- as.data.frame(matrix(sample(1:100, 100), nrow = 10, dimnames = list(paste("row", 1:10, sep=""), paste("col", 1:10, sep=""))))
x
dfRowName(x, name = "ColumnWithRowNames")
}
