% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity_to_distance.R
\name{dissimilarity_to_distance_importance_plot}
\alias{dissimilarity_to_distance_importance_plot}
\title{Display species importance plot based on a weighted Euclidean distance transformation of dissimilarities.}
\usage{
dissimilarity_to_distance_importance_plot(
  x,
  scaled_weights = TRUE,
  mark_removed = TRUE
)
}
\arguments{
\item{x}{Result of \code{\link{dissimilarity_to_distance}} function}

\item{scaled_weights}{Logical; if TRUE, scaled weights will be plotted}

\item{mark_removed}{Logical; if TRUE, species that were removed from distance will be highlighted with a separate symbol on a plot}
}
\value{
ggplot-object.
}
\description{
Display species importance plot based on a weighted Euclidean distance transformation of dissimilarities.
}
\details{
Species importance measures the contribution of individual species to the overall dissimilarity between the samples.
}
\examples{
library(vegan); library(smacof)
data(varespec)

# Estimate weighted Euclidean distance
bc_to_eucl <- dissimilarity_to_distance(varespec, dist_type = "bray", drop_species = T, importance_percentile = 0.05, show_plot = F)

# Plot species weigths
dissimilarity_to_distance_importance_plot(bc_to_eucl)

}
\seealso{
\code{\link{dissimilarity_to_distance}}
}
