% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2list.R
\name{dist2list}
\alias{dist2list}
\title{Convert distance matrix to data frame}
\usage{
dist2list(dist, tri = TRUE)
}
\arguments{
\item{dist}{Distance matrix (object of class 'dist')}

\item{tri}{Logical, if TRUE - only lower triangular part of dist will be returned}
}
\value{
Data frame
}
\description{
This function takes a distance matrix (of class 'dist') and transforms it to a data.frame, where each row represents a single pairwise comparison.
}
\examples{
# Generate dummy data
datt <- dist(rnorm(10))
attr(datt, "Labels") <- paste("obj", 1:10, sep="")

# Convert dist to data frame
head( dist2list(datt)  )
dim( dist2list(datt, tri = T)  ) # only lower-triangular part of dist
dim( dist2list(datt, tri = F)  ) # full distance matrix

}
\seealso{
\code{\link[spaa]{dist2list}} from spaa package.
}
\author{
The code is based on \code{\link[spaa]{dist2list}} by Jinlong Zhang.
}
