% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig_perc.R
\name{eig_perc}
\alias{eig_perc}
\title{Estimate the proportion of explained variance from the eigenvalues.}
\usage{
eig_perc(eig, positive = T, plot = F, ...)
}
\arguments{
\item{eig}{Numeric vector of eigenvalues (e.g., from PCA or PCoA)}

\item{positive}{Logical, preserve only positive eigenvalues (if TRUE negative eigenvalues will be removed)}

\item{plot}{Logical, show a scree plot}

\item{...}{Additinal arguments will be passed to a plot function}
}
\value{
Numeric vector of the amounts of total variance captured by principal components or principal coordinates.
}
\details{
Negative eigenvalue could be generated in the analysis of semi- or non-metric dissimilarity measures.
}
\examples{
pco <- cmdscale(d = eurodist, k = 10, eig = TRUE)
eig_perc(pco$eig, positive = T, plot = T)

}
