% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_CIGAR.R
\name{expand_CIGAR}
\alias{expand_CIGAR}
\title{Expand CIGAR Strings}
\usage{
expand_CIGAR(cigar)
}
\arguments{
\item{cigar}{A character vector containing CIGAR strings to be expanded.}
}
\value{
A character vector with the expanded CIGAR strings.
}
\description{
This function expands compact CIGAR (Compact Idiosyncratic Gapped Alignment Report) strings
into a full-length sequence of CIGAR codes ("operations").
}
\details{
CIGAR is a compact representation of an alignment,
commonly utilized in the SAM file format.
A CIGAR string consists of `<N><X>` pairs, such as "5M2D",
where `X` represents an operation designated by an uppercase letter (refer to the table below for details),
and the integer `N` indicates the count of consecutive operations.
The function, as currently implemented, accommodates CIGAR strings
in the style of USEARCH/VSEARCH, wherein the integer can be omitted when it equals 1.

The following table describes the CIGAR string operation codes:
\describe{
  \item{\code{Operation}}{\code{Description}}
  \item{\code{M}}{Match or mismatch}
  \item{\code{D}}{Deletion (gap in the target sequence).}
  \item{\code{I}}{Insertion (gap in the query sequence).}
  \item{\code{S}}{Segment of the query sequence that does not appear in the alignment, used with soft clipping.}
  \item{\code{H}}{Segment of the query sequence that does not appear in the alignment, used with hard clipping.}
  \item{\code{X}}{Mismatch (supported, but USEARCH/VSEARCH does not generate this code).}
}
}
\examples{
expand_CIGAR("5MI2D")
expand_CIGAR(c("5MI2D", "3M2D3M"))
expand_CIGAR(c("5MI2D", "*", "3M2D3M", NA))  # handles centroids and missing data

}
\seealso{
\code{\link[GenomicAlignments]{`cigar-utils`}}
}
