% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_in_parentheses.R
\name{extract_in_parentheses}
\alias{extract_in_parentheses}
\title{Extract information inside the parentheses from a character string.}
\usage{
extract_in_parentheses(x)
}
\arguments{
\item{x}{Character string}
}
\value{
Part(s) of the input character string that was constrained in the parentheses.
}
\description{
Extract information inside the parentheses from a character string.
}
\examples{
extract_in_parentheses("abc (11\%) xyz")
extract_in_parentheses("Lorem ipsum (dolor sit amet), consectetur adipiscing elit, (sed do eiusmod) tempor incididunt ut (labore et dolore) magna aliqua")

}
\references{
This function is based on the answer from https://stackoverflow.com/questions/8613237/extract-info-inside-all-parenthesis-in-r
}
