% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cazy.R
\name{filter_cazy}
\alias{filter_cazy}
\title{Filter CAZy table}
\usage{
filter_cazy(x, min_abund = 15, min_occurence = 10)
}
\arguments{
\item{x}{Data frame with gene counts (every column represents one CAZy annotation and every row represents one organism)}

\item{min_abund}{The minimum CAZy count across all organisms to be retained (default, 15)}

\item{min_occurence}{The minimum number of organisms with a given CAZy annotation to be retained (default, 10)}
}
\value{
Subsetted data frame
}
\description{
This functions filters CAZy annotations based on the total abundance and occurence of CAZy genes across organisms. In other words, it removes CAZy annotations with low abundance.
}
\examples{

}
