% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_taxonomic_rank_DT.R
\name{get_max_taxonomic_rank_DT}
\alias{get_max_taxonomic_rank_DT}
\title{Determine the lowest level of taxonomic classification}
\usage{
get_max_taxonomic_rank_DT(x, return_rank_only = FALSE)
}
\arguments{
\item{x}{Either a phyloseq object, or a data frame with columns as taxonomic ranks and rows as entries (e.g., OTUs). Columns in the data frame should be ordered from the highest level of classification (e.g., Kingdom) to the lowest level (e.g., Species), missing data are coded as NA}

\item{return_rank_only}{Logical, if TRUE only name of the taxonomic rank will be returned}
}
\value{
Data frame with taxonomy and additional column containing the name of the lowest level of taxonomic classification. Alternatively, if 'return_rank_only = TRUE', a vector of the lowest taxonomic ranks for each OTU.
}
\description{
Determine the lowest level of taxonomic classification
}
\details{
This function will find the last non-NA column in the taxonomy table and return
}
\examples{
data(GlobalPatterns)

# phyloseq-class as input
taxx <- get_max_taxonomic_rank(GlobalPatterns)
summary(taxx$RankName)

# data frame as input
taxtbl <- as.data.frame(tax_table(GlobalPatterns))
taxx <- get_max_taxonomic_rank(taxtbl)
summary(taxx$RankName)

}
