% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_utax_taxonomy.R
\name{make_utax_taxonomy}
\alias{make_utax_taxonomy}
\title{Prepare taxonomy annotations in SINTAX or UTAX style.}
\usage{
make_utax_taxonomy(x)
}
\arguments{
\item{x}{Data frame with one row (columns are ordered taxonomy levels)}
}
\value{
Character string.
}
\description{
This function takes taxonomy table and returns character string in USEARCH-style that can be used with SINTAX or UTAX taxonomy classificators.
}
\details{
Missing values are allowed and should be coded as 'NA'.
}
\examples{
# Create dummy data (for one species)
datt <- data.frame(
  Kingdom = "Fungi", Phylum = "Basidiomycota", Class = "Agaricomycetes",
  Order = "Agaricales", Family = "Marasmiaceae", Genus = "Marasmius", Species = "Marasmius_alliaceus",
  stringsAsFactors = F)

make_utax_taxonomy(datt)

}
\seealso{
http://drive5.com/usearch/manual/tax_annot.html
}
