% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_utax_taxonomy.R
\name{make_utax_taxonomy_batch}
\alias{make_utax_taxonomy_batch}
\title{Prepare taxonomy annotations in SINTAX or UTAX style for multiple species.}
\usage{
make_utax_taxonomy_batch(x, as_vector = T, ...)
}
\arguments{
\item{x}{Data frame (columns are ordered taxonomy levels, rows are species or OTUs)}

\item{as_vector}{Logical, if TRUE (default) result will be returned as vector. Otherwise, result is a data frame}

\item{...}{Additional arguments may be passed to \code{\link{make_utax_taxonomy}}}
}
\value{
Vector or a data frame with taxonomy strings in USEARCH format.
}
\examples{
# Create dummy data (each row = species)
datt <- data.frame(
  Kingdom = rep(LETTERS[1:3], each = 12),
  Phylum = rep(LETTERS[6:11], each = 6),
  Class = rep(letters[1:12], each = 3),
  Order = do.call(paste0, expand.grid(letters, letters, stringsAsFactors = F))[1:36],
  stringsAsFactors = F
)

head(datt)  # Taxonomy table
head(make_utax_taxonomy_batch(datt))  # Taxonomy strings for USEARCH classifiers

}
\seealso{
http://drive5.com/usearch/manual/tax_annot.html
}
