% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_mult_dist.R
\name{mult_dist_average}
\alias{mult_dist_average}
\title{Average multiple distance matrices.}
\usage{
mult_dist_average(dlist)
}
\arguments{
\item{dlist}{List of distance matrices (class 'dist' or 'matrix')}
}
\value{
Distance matrix of the same class as input matrices ('dist' or 'matrix') with averaged values.
}
\description{
This function can be used in order to avergage beta diversity over multiple rarefaction iterations.
}
\examples{
library(plyr)
# Generate dummy data (list with 3 5x5 matrices)
ddd <- rlply(.n = 3, .expr = function(){ as.dist( matrix(sample(1:1000, size = 25, replace = T), ncol = 5)) })
ddd

# Average matrices
mult_dist_average(ddd)

}
