% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_amplicon_table.R
\name{parse_amplicon_table}
\alias{parse_amplicon_table}
\title{Parse amplicon table}
\usage{
parse_amplicon_table(amp = otu.tab, tax = blast, tax_type = "BLAST",
  add.tax.to.data = TRUE, tax.db)
}
\arguments{
\item{amp}{File name with OTU table (generated with VSEARCH or USEARCH)}

\item{tax}{File name with OTU taxonomic annotations}

\item{tax_type}{Type of taxonomic annotation ("BLAST", "SINA", "STAMPA")}

\item{add.tax.to.data}{logical, add taxonomic annotations to the OTU table?}

\item{tax.db}{SILVA database with taxonomy strings and corresponding tax ranks (e.g. tax_slv_ssu_128.txt)}
}
\value{
list
}
\examples{

}
