% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fastq_eestats2.R
\name{parse_fastq_eestats2}
\alias{parse_fastq_eestats2}
\title{Load USEARCH fastq_eestats2 report}
\usage{
parse_fastq_eestats2(file, long = FALSE)
}
\arguments{
\item{file}{the name of the input file.}

\item{long}{a logical value indicating wheather the results should be returned in long format.}
}
\value{
data frame
}
\description{
Load USEARCH fastq_eestats2 report showing how many reads will pass an expected error filter at different length thresholds.
}
\details{
This function depends on plyr and reshape2 packages.
Additional details on 'fastq_eestats2' function can be found on http://drive5.com/usearch/manual/cmd_fastq_eestats2.html
}
\examples{
## Run USEARCH
# usearch -fastq_eestats2 reads.fq -output eestats2.txt -length_cutoffs 200,300,10
parse_fastq_eestats2("eestats2.txt", long = F)
parse_fastq_eestats2("eestats2.txt", long = T)

library(ggplot2)
ee <- parse_fastq_eestats2("eestats2.txt", long = T)
ggplot(data = ee, aes(x = Length, y = NumberOfReads, group = MaxEE, color = as.factor(MaxEE))) + 
  geom_point(size = 2) + 
  geom_line() + 
  theme_classic()

}
