% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_funguild.R
\name{parse_funguild}
\alias{parse_funguild}
\title{Parse FUNGuild database}
\usage{
parse_funguild(url = "http://www.stbates.org/funguild_db.php",
  tax_name = TRUE)
}
\arguments{
\item{url}{Funguild databse link}

\item{tax_name}{Logical, change numeral taxonomic coding with the corresponding taxonomic levels}
}
\value{
Data frame containing the entire FUNGuild database.
}
\description{
FUNGuild = Fungi + fUNctional + Guild
}
\examples{
fg <- parse_funguild()

}
\references{
Nguyen NH, Song Z, Bates ST, Branco S, Tedersoo L, Menke J, Schilling JS, Kennedy PG. 2016. FUNGuild: An open annotation tool for parsing fungal community datasets by ecological guild. Fungal Ecology 20:241–248.
https://github.com/UMNFuN/FUNGuild
}
