% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_silva.R
\name{parse_silva_tax}
\alias{parse_silva_tax}
\title{Split single SILVA taxonomic annotation to taxonomic ranks.}
\usage{
parse_silva_tax(x, db)
}
\arguments{
\item{x}{Character string with taxonomic annotation from BLAST.}

\item{db}{Data frame with rank designations for all taxonomic paths in the SILVA taxonomy.}
}
\value{
Named charcter vector with taxonomic annotation (names = rank designations).
}
\details{
This function will split the taxonomic annotation based on the SILVA database and assign a corresponding taxonomic rank to the obtained parts.
}
\examples{
# Download taxonomic rank designations for all taxonomic paths used in the SILVA taxonomies
tax.db <- read.delim("https://www.arb-silva.de/fileadmin/silva_databases/release_128/Exports/taxonomy/tax_slv_ssu_128.txt", header = F, stringsAsFactors = F)
colnames(tax.db) <- c("path", "taxid", "rank", "remark", "release")

x <- "Eukaryota;SAR;Alveolata;Ciliophora;Intramacronucleata;Conthreep;Oligohymenophorea;Peritrichia;Telotrochidium;uncultured eukaryote"
parse_silva_tax(x, tax.db)

}
\seealso{
\code{\link{parse_silva_tax_batch}}
}
