% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_silva.R
\name{parse_silva_tax_batch}
\alias{parse_silva_tax_batch}
\title{Split vector of SILVA taxonomic annotations and prepare taxonomy table.}
\usage{
parse_silva_tax_batch(x, db)
}
\arguments{
\item{x}{Vector of character strings with taxonomic annotation (assigned by BLAST)}

\item{db}{Data frame with rank designations for all taxonomic paths in the SILVA taxonomy}
}
\value{
Data frame with OTUs or species as rows and their taxonomic ranks as columns.
}
\examples{
# Download taxonomic rank designations for all taxonomic paths used in the SILVA taxonomies
tax.db <- read.delim("https://www.arb-silva.de/fileadmin/silva_databases/release_128/Exports/taxonomy/tax_slv_ssu_128.txt", header = F, stringsAsFactors = F)
colnames(tax.db) <- c("path", "taxid", "rank", "remark", "release")

x <- c("Eukaryota;Opisthokonta;Holozoa;Metazoa (Animalia);Eumetazoa;Bilateria;Platyhelminthes;Catenulida;Catenulidae;Paracatenula galateia",
"Eukaryota;SAR;Rhizaria;Cercozoa;Imbricatea;Silicofilosea;Euglyphida;Assulinidae;Assulina;Assulina muscorum",
"Eukaryota;Opisthokonta;Nucletmycea;Fungi;Cryptomycota;LKM11;uncultured eukaryote",
"Eukaryota;Opisthokonta;Nucletmycea;Fungi;Dikarya;Basidiomycota;Ustilaginomycotina;Exobasidiomycetes;Malasseziales;Incertae Sedis;Malassezia;Malassezia restricta CBS 7877",
"Eukaryota;Opisthokonta;Nucletmycea;Fungi;Cryptomycota;LKM11;uncultured fungus",
"Eukaryota;SAR;Alveolata;Ciliophora;Intramacronucleata;Conthreep;Oligohymenophorea;Peritrichia;Telotrochidium;uncultured eukaryote",
"No blast hit", "Eukaryota;SAR;Alveolata;Ciliophora;Intramacronucleata;Armophorea;Armophorida;Metopus;Metopus striatus",
"Eukaryota;Opisthokonta;Holozoa;Metazoa (Animalia);Eumetazoa;Bilateria;Rotifera;Monogononta;Ploimida;Lepadella patella",
"Eukaryota;SAR;Rhizaria;Cercozoa;Novel Clade 10;uncultured eukaryote"
)
parse_silva_tax_batch(x, tax.db)

}
\seealso{
\code{\link{parse_silva_tax}}
}
