% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_swarms.R
\name{parse_swarms}
\alias{parse_swarms}
\title{Parse SWARM files}
\usage{
parse_swarms(x, otu_delimiter = " ")
}
\arguments{
\item{x}{Input file name (usually with "swarm" extension)}

\item{otu_delimiter}{Character string delimiting sequence IDs within a file line (default, whitespace)}
}
\value{
data frame with two columns: OTU name and sequences names that belongs to this OTU.
}
\description{
This function reads the results of SWARM clustering (*.swarms file).
}
\details{
SWARM is a single-linkage clustering method with iterative growth process and the use of sequence abundance values to delineate OTUs.
SWARM properly delineates large OTUs (high recall), and can distinguish OTUs with as little as two differences between their centers (high precision).
}
\examples{
parse_swarms("tst_1f.swarms")


}
\seealso{
Mahé F, Rognes T, Quince C, de Vargas C, Dunthorn M. (2015) Swarm v2: highly-scalable and high-resolution amplicon clustering. PeerJ 3:e1420 doi: 10.7717/peerj.1420
https://github.com/torognes/swarm
https://github.com/frederic-mahe/swarm/wiki/Fred's-metabarcoding-pipeline
}
