% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phred_scores.R
\name{phred_inverse}
\alias{phred_inverse}
\title{Estimate probability of an error based on the Phred-score.}
\usage{
phred_inverse(q)
}
\arguments{
\item{q}{= Phred Quality Score}
}
\value{
Probability that the base is called wrong.
}
\description{
Estimate probability of an error based on the Phred-score.
}
\examples{
phred_inverse()
phred_inverse(30)
data.frame(Q = 2:42, ProbError = phred_inverse(2:42))

}
\seealso{
\code{\link{phred_score}}
}
