% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phred_scores.R
\name{phred_score}
\alias{phred_score}
\title{Estimate Phred Quality Score score from probability of an error}
\usage{
phred_score(p)
}
\arguments{
\item{p}{Probability that the base is called wrong.}
}
\value{
Phred Q-Score
}
\examples{
phred_score(c(0.001, 0.01, 0.05, 0.1))

}
\seealso{
\code{\link{phred_inverse}}
}
