% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_add_max_tax_rank.R
\name{phyloseq_add_max_tax_rank}
\alias{phyloseq_add_max_tax_rank}
\title{Add the lowest level of taxonomic classification to the taxonomy table of phyloseq-class object.}
\usage{
phyloseq_add_max_tax_rank(physeq, abbreviate = F, ...)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{abbreviate}{Logical; if TRUE, lowest taxon name will be abbreviated (useful for plotting). Default is FALSE}

\item{...}{Additional arguments may be passed to \code{\link{abbreviate_taxa_names}}}
}
\value{
Phyloseq object with modified taxonomy table.
}
\examples{

}
