% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_effort.R
\name{phyloseq_effort_div_rangeplot}
\alias{phyloseq_effort_div_rangeplot}
\title{Vizualize OTU diversity and richness for a particular sequencing depth or sample coverage ranges}
\usage{
phyloseq_effort_div_rangeplot(
  physeq,
  range,
  range_type = "coverage",
  variable = NULL,
  yvar = "seqdepth",
  interval = "quartiles",
  show_plot = TRUE,
  justDF = FALSE,
  progress = "text",
  ...
)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{range}{Numeric vector of coverage levels or sequence depths to test, e.g., seq(from = 100, to = 1000, by = 100)}

\item{range_type}{"coverage" or "seqdepth"}

\item{variable}{Groupping variable name (contained in \code{\link[phyloseq]{sample_data}}); default, NULL}

\item{yvar}{Which dependent variable to show ("coverage", "seqdepth", "Q1", "Q2", "Q3").
For range_type = "seqdepth": "coverage", "Q0", "Q1", "Q2".
For range_type = "coverage": "seqdepth", "Q0", "Q1", "Q2"}

\item{interval}{Which interval to show on the group-averaged data ("CI", "quartiles", or NULL)}

\item{show_plot}{Logical; if TRUE, show the resulting plot}

\item{justDF}{Logical; if TRUE, just return the resulting table}

\item{progress}{Show progress bar ("text" or "none")}

\item{...}{Additional arguments may be passed to \code{\link{phyloseq_effort_div}}}
}
\value{
ggplot object (if show_plot = TRUE) or a data.frame (if justDF = TRUE)
}
\description{
This function helps to estimate OTU diversity and sample coverage for a particular sequencing depth or
to estimate the required sequencing depth and OTU diversity for a particular sample coverage
}
\details{
Diversity estimation based on the analytical approach for coverage-based rarefaction proposed by Chao and Jost (2012).
}
\examples{
data("esophagus")
# Show sequencing depth required to reach a particular sample coverage
phyloseq_effort_div_rangeplot(esophagus, range = seq(from = 0.8, to = 0.97, by = 0.02),
    range_type = "coverage", yvar = "seqdepth", interval = "quartiles")

# Show OTU richness (Q = 0) achieved by different sequencing efforts
phyloseq_effort_div_rangeplot(esophagus, range = seq(from = 100, to = 500, by = 20),
    range_type = "seqdepth", yvar = "Q0", interval = "quartiles")


data("GlobalPatterns")
# Subset data
GP <- subset_samples(GlobalPatterns, SampleType \%in\% c("Soil", "Feces", "Skin"))
GP <- phyloseq_filter_top_taxa(GP, perc = 20)

# Show sequencing depth required to reach a particular sample coverage level for different sample groups
phyloseq_effort_div_rangeplot(GP, range = seq(from = 0.8, to = 0.97, by = 0.02),
  variable = "SampleType", range_type = "coverage", yvar = "seqdepth", interval = "quartiles")

# Show OTU richness (Q = 0) achieved by different sequencing efforts for different sample groups
phyloseq_effort_div_rangeplot(GP, range = seq(from = 100, to = 500, by = 20),
   variable = "SampleType", range_type = "seqdepth", yvar = "Q0")

}
