% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_extract_otus.R
\name{phyloseq_extract_shared_otus}
\alias{phyloseq_extract_shared_otus}
\title{Extract common species (OTUs) between samples.}
\usage{
phyloseq_extract_shared_otus(x, samp_names = sample_names(x))
}
\arguments{
\item{x}{A phyloseq-class object}

\item{samp_names}{Character vector with sample names (default, all samples)}
}
\value{
Phyloseq object with the subset of data.
}
\description{
This function will subset phyloseq object to the OTUs that are present in all samples. By default all samples will be taken into account, otherwise it's possible to take a subset of samples.
}
\examples{
data(esophagus)

# Extract OTUs that are present in all samples
ps <- phyloseq_extract_shared_otus(esophagus, samp_names = sample_names(esophagus))
ps
otu_table(ps)

# Extract shared OTUs between two samples
ps2 <- phyloseq_extract_shared_otus(esophagus, samp_names = c("B", "C"))
ps2
otu_table(ps2)

}
