% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_filter.R
\name{phyloseq_filter_sample_wise_abund_trim}
\alias{phyloseq_filter_sample_wise_abund_trim}
\title{Filter rare OTUs based on minimum abundance threshold.}
\usage{
phyloseq_filter_sample_wise_abund_trim(physeq, minabund = 10,
  rm_zero_OTUs = TRUE)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{minabund}{Abundance threshold (default, 10)}

\item{rm_zero_OTUs}{Logical, remove OTUs with zero total abundance}
}
\value{
Phyloseq object with a filtered data.
}
\description{
This function performs sample-wise OTU abundance trimming.
}
\details{
OTUs can be considered as rare if they comprise fewer than X (e.g., 10) sequences within a sample. This function is intented to censore OTU abundance (unsing an arbitrary threshold) on a sample-wise basis.
}
\examples{
# Load data
data(GlobalPatterns)

# Trim GlobalPatterns data (19216 taxa)
GP <- phyloseq_filter_sample_wise_abund_trim(GlobalPatterns) # 10605 taxa

# Compare raw and trimmed data
phyloseq_compare(GlobalPatterns, GP, cols = c("GlobalPatterns", "Trimmed GlobalPatterns"))

}
