% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_filter.R
\name{phyloseq_filter_taxa_rel_abund}
\alias{phyloseq_filter_taxa_rel_abund}
\title{Remove taxa with small mean relative abundance.}
\usage{
phyloseq_filter_taxa_rel_abund(physeq, frac = 1e-04)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{frac}{The minimum cutoff for the relative OTU abundance}
}
\value{
Phyloseq object with a subset of taxa.
}
\details{
This function searches for taxa with small mean relative abundance and removes them. Result will be returned with original counts in the abundance table.
}
\examples{
data("esophagus")
phyloseq_filter_taxa_rel_abund(esophagus, frac = 0.01)
phyloseq_filter_taxa_rel_abund(esophagus, frac = 0.1)

}
