% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_filter.R
\name{phyloseq_filter_taxa_tot_fraction}
\alias{phyloseq_filter_taxa_tot_fraction}
\title{Remove taxa with abundance less then a certain fraction of total abundance.}
\usage{
phyloseq_filter_taxa_tot_fraction(physeq, frac = 0.01)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{frac}{The minimum cutoff for the OTU abundance in the table. This number is a fraction, not a percent.}
}
\value{
Phyloseq object with a subset of taxa.
}
\details{
If frac = 0.0001, this will retain all OTU's that have at least a 0.01% total abundance in the OTU table.
If you wanted to retain OTUs with at least 1% total abundance, you must specify, 0.01.
}
\examples{
data("esophagus")
phyloseq_filter_taxa_tot_fraction(esophagus, frac = 0.01)

}
\seealso{
http://qiime.org/scripts/filter_otus_from_otu_table.html
}
