% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_filter.R
\name{phyloseq_filter_top_taxa}
\alias{phyloseq_filter_top_taxa}
\title{Extract the most abundant taxa.}
\usage{
phyloseq_filter_top_taxa(physeq, perc = 10, n = NULL)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{perc}{Percentage of the most abundant taxa to retain}

\item{n}{Number of the most abundant taxa to retain (this argument will override perc argument)}
}
\value{
Phyloseq object with a filtered data.
}
\description{
Extract the most abundant taxa.
}
