% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_filter.R
\name{phyloseq_filter_top_taxa_range}
\alias{phyloseq_filter_top_taxa_range}
\title{Check the range of the top-taxa filtering values to determine the optimal threshold.}
\usage{
phyloseq_filter_top_taxa_range(physeq, show_plot = TRUE)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{show_plot}{Logical; if TRUE, shows the plot on screen}
}
\value{
ggplot-object.
}
\description{
This function performs taxa filtering by retaining the most abundant taxa.
A range of abundance percentages (5 - 95\%) will be explored.
}
\examples{

}
