% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_merge_samples.R
\name{phyloseq_merge_samples}
\alias{phyloseq_merge_samples}
\title{Merge samples by name}
\usage{
phyloseq_merge_samples(phys, samples_to_merge, new_id = NULL)
}
\arguments{
\item{phys}{A phyloseq-class object}

\item{samples_to_merge}{Character vector of sample names that should be merged}

\item{new_id}{Name of the new sample}
}
\value{
Phyloseq-object.
}
\description{
This function will merge samples (specified by their names) in a phyloseq-object.
}
\details{
Sample metadata of the first ID from `samples_to_merge` will be used as a representative for a group.
}
