% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_mult_raref.R
\name{phyloseq_mult_raref}
\alias{phyloseq_mult_raref}
\title{Multiple rarefaction of phyloseq-object}
\usage{
phyloseq_mult_raref(
  x,
  SampSize = NULL,
  MinSizeThreshold = NULL,
  iter = 1000,
  replace = F,
  multithread = F,
  seeds = NULL,
  ...
)
}
\arguments{
\item{x}{A phyloseq-class object}

\item{SampSize}{Rarefaction depth (number of reads to sample)}

\item{MinSizeThreshold}{Remove samples with number of reads less than this threshold}

\item{iter}{Number of rarefication iterations}

\item{replace}{Logical, whether to sample with replacement (TRUE) or without replacement (FALSE, default)}

\item{multithread}{Logical or integer; if TRUE, attempts to run the function on multiple cores; integer defines the number of cores to use (if it is set to TRUE, all cores will be used)}

\item{seeds}{Integer vector used for the reproducible random subsampling (should be of the same length as the number of iterations)}

\item{...}{Additional arguments will be passed to \code{\link[phyloseq]{rarefy_even_depth}}}
}
\value{
List of rarefied phyloseq-objects.
}
\description{
This function will resample an OTU table such that all samples have the same library size. Resampling will be performed multiple times.
}
\details{
If the sample size ('SampSize') is not specified, rarefaction will be made for the depth equeal to 0.9 * minimal observed sample size.
By default, sampling is performed without replacement ('replace = FALSE'), which differs from the default behaviour of \code{\link[phyloseq]{rarefy_even_depth}}.
}
\examples{
library(phyloseq)
data("esophagus")
eso_raref <- phyloseq_mult_raref(esophagus, iter = 10)

# Discard samples with number of reads < 210 and perform resampling with replacement
eso_raref_t <- phyloseq_mult_raref(esophagus, replace = T, MinSizeThreshold = 210, SampSize = 210, iter = 10)
sample_sums(esophagus)
sample_sums(eso_raref_t[[1]])

# Do not remove OTUs from the dataset that are no longer observed in any sample (have a count of zero in every sample)
phyloseq_mult_raref(esophagus, trimOTUs = F, replace = T, MinSizeThreshold = 210, SampSize = 210, iter = 10)

}
\seealso{
\code{\link[phyloseq]{rarefy_even_depth}}
}
