% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_mult_raref_avg.R
\name{phyloseq_mult_raref_avg}
\alias{phyloseq_mult_raref_avg}
\title{Perform rarefaction and average relative OTU abundance}
\usage{
phyloseq_mult_raref_avg(physeq, SampSize = NULL, iter = 1000,
  parallel = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{...}{passed to phyloseq_mult_raref}
}
\examples{

}
