% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_mult_raref_div.R
\name{phyloseq_mult_raref_div}
\alias{phyloseq_mult_raref_div}
\title{Average diversity estimates across multiple rarefaction iterations.}
\usage{
phyloseq_mult_raref_div(physeq, SampSize = min(sample_sums(physeq)),
  iter = 1000, divindex = c("Observed", "Shannon"), parallel = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{SampSize}{Rarefaction depth (number of reads to sample)}

\item{iter}{Number of rarefication iterations}

\item{divindex}{Alpha-diversity measures to estimate (for the supported indices see \code{\link[phyloseq]{estimate_richness}})}

\item{parallel}{Logical; if TRUE, attempts to run the function on multiple cores}

\item{verbose}{Logical; if TRUE, progress messages from the function will be printed}

\item{...}{Additional argument may be passed to \code{\link{phyloseq_mult_raref}}}
}
\value{
A data.frame of the richness estimates.
}
\description{
This function performs multiple rarefaction and estimates average diversity (e.g., Simpson or Shannon indices) for each sample.
}
