% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_ntaxa_by_tax.R
\name{phyloseq_ntaxa_by_tax}
\alias{phyloseq_ntaxa_by_tax}
\title{Count number of OTUs by taxonomic rank for each sample.}
\usage{
phyloseq_ntaxa_by_tax(
  physeq,
  TaxRank = "Phylum",
  relative = F,
  add_meta_data = T
)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{TaxRank}{Name of the taxonomic rank}

\item{relative}{Logical, return relative number of OTUs}

\item{add_meta_data}{Logical, add sample metadata to the resulting table}
}
\value{
Data frame with OTU counts (columns: Sample, TaxRank, N.OTU + optional sample metadata).
}
\description{
Count number of OTUs by taxonomic rank for each sample.
}
\details{
This function will split phyloseq object by the specified taxonomic rank
}
\examples{
data(GlobalPatterns)
phylum_counts <- phyloseq_ntaxa_by_tax(GlobalPatterns, TaxRank = "Phylum")
head(phylum_counts)

}
