% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_extract_otus.R
\name{phyloseq_otu_appearance}
\alias{phyloseq_otu_appearance}
\title{Determine which OTUs appeared or disappeared in comparison with reference sample group.}
\usage{
phyloseq_otu_appearance(phys, ref_level)
}
\arguments{
\item{phys}{A phyloseq-class object}

\item{ref_level}{Sample name(s) relative to which appearance or disappearance of OTU will be measured}
}
\value{
List with OTU names that appeared, disappeared or remained.
}
\description{
Determine which OTUs appeared or disappeared in comparison with reference sample group.
}
