% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_prevalence_plot.R
\name{phyloseq_prevalence_plot}
\alias{phyloseq_prevalence_plot}
\title{Prevalence plot (total OTU abundance vs OTU prevalence).}
\usage{
phyloseq_prevalence_plot(physeq, prev.trh = NULL, taxcolor = NULL,
  facet = FALSE, point_alpha = 0.7, showplot = T)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{prev.trh}{Add horizontal line with prevalence threshold (default is NULL, but  0.05 = 5\% of samples)}

\item{taxcolor}{Taxonomy rank for coloring the points (e.g. "Phylum")}

\item{facet}{Logical, split to separate panels by taxonomy rank used for coloring the points}

\item{point_alpha}{Point transparency value}

\item{showplot}{Logical, show plot on screen}
}
\value{
Plot of class 'ggplot'.
}
\description{
This function will plot total OTU abundance vs the fraction of samples in which an OTU is observed.
}
\examples{
data(GlobalPatterns)
# Subset data
GP <- subset_taxa(GlobalPatterns, Phylum \%in\% c("Acidobacteria", "Actinobacteria", "Firmicutes", "Verrucomicrobia"))
phyloseq_prevalence_plot(GP, taxcolor = "Phylum", facet = TRUE, point_alpha = 0.5, prev.trh = 0.05)

}
\seealso{
\code{\link{phyloseq_filter_prevalence}}
}
