% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_randomize.R
\name{phyloseq_randomize}
\alias{phyloseq_randomize}
\title{Randomize abundance table and phylogeny in phyloseq objects (for null model testing and simulation).}
\usage{
phyloseq_randomize(physeq, null_model = "phylogeny.pool", verbose = T, ...)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{null_model}{Character string defining the null model (for the description of supported models see \code{\link[picante]{randomizeMatrix}})}

\item{verbose}{Logical; if TRUE, additional information messages will be displayed}

\item{...}{Additional arguments may be passed to \code{\link[picante]{randomizeMatrix}}}
}
\value{
A phyloseq-class object with randomized abundance table and/or phylogeny.
}
\details{
Currently only null models from picante package are implemented.
}
\examples{

}
\seealso{
\code{\link[picante]{randomizeMatrix}}, \code{\link[vegan]{commsim}}
}
