% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_raref_rtk.R
\name{phyloseq_raref_rtk}
\alias{phyloseq_raref_rtk}
\title{RTK-based rarefaction}
\usage{
phyloseq_raref_rtk(
  physeq,
  SampSize = NULL,
  MinSizeThreshold = NULL,
  iter = 1000,
  trimOTUs = TRUE,
  ...
)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{SampSize}{Rarefaction depth (number of reads to sample)}

\item{MinSizeThreshold}{Remove samples with number of reads less than this threshold}

\item{iter}{Number of rarefication iterations}

\item{trimOTUs}{Logical, if TRUE (default), OTUs that have a count of zero in every sample will be removed}

\item{...}{Additional arguments will be passed to \code{\link[rtk]{rtk}}}
}
\value{
List of rarefied phyloseq-objects.
}
\description{
RTK-based rarefaction
}
\examples{
data("esophagus")
eso_raref <- phyloseq_raref_rtk(esophagus, iter = 10)

# Benchmark
system.time( GPr <- phyloseq_raref_rtk(GlobalPatterns, SampSize = 5000, iter = 10, trimOTUs = T) )
system.time( GPr <- phyloseq_mult_raref(GlobalPatterns, SampSize = 5000, iter = 10, trimOTUs = T) )

}
