% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_rename_with_tax.R
\name{phyloseq_rename_with_tax}
\alias{phyloseq_rename_with_tax}
\title{Rename phyloseq OTUs or species with taxonomy ranks.}
\usage{
phyloseq_rename_with_tax(physeq, taxrank = "Family")
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{taxrank}{Character, which taxonomy rank to use for renaming}
}
\value{
A phyloseq-class object with renamed OTUs.
}
\description{
Rename phyloseq OTUs or species with taxonomy ranks.
}
\examples{
# Load data
data("GlobalPatterns")

# Agglomerate taxa of the same phylum
GP <- tax_glom(GlobalPatterns, taxrank="Phylum")
taxa_names(GP)

# Assign phylum name to the agglomerated taxa
GP <- phyloseq_rename_with_tax(GP, taxrank="Phylum")
taxa_names(GP)

}
