% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_transform.R
\name{phyloseq_replace_zero}
\alias{phyloseq_replace_zero}
\title{Replace zeros in OTU abundance}
\usage{
phyloseq_replace_zero(physeq, method = "pseudocount", pseudocount = 0.65)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{method}{Which method to apply for zero replacement ("pseudocount")}

\item{pseudocount}{Constant value}
}
\value{
Phyloseq object with transformed counts in OTU table.
}
\description{
Replace zeros in OTU abundance
}
