% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_filter.R
\name{phyloseq_richness_filter}
\alias{phyloseq_richness_filter}
\title{Remove samples from phyloseq object that have less than n taxa}
\usage{
phyloseq_richness_filter(physeq, mintaxa = 10)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{mintaxa}{Minimum number of taxa that should be present in a sample (default, 10)}
}
\value{
Trimmed phyloseq object (All samples will have >= N taxa)
}
\description{
Remove samples from phyloseq object that have less than n taxa
}
\examples{
data("esophagus")
esophagus
phyloseq_richness_filter(esophagus, mintaxa = 30)
phyloseq_richness_filter(esophagus, mintaxa = 100)

}
