% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_rm_na_tax.R
\name{phyloseq_rm_na_tax}
\alias{phyloseq_rm_na_tax}
\title{Remove unused taxonomy levels from phyloseq-object.}
\usage{
phyloseq_rm_na_tax(physeq)
}
\arguments{
\item{physeq}{A phyloseq-class object.}
}
\value{
Phyloseq object.
}
\details{
This function is particularly useful after \code{\link{tax_glom}}.

'phyloseq_rm_na_tax' will remove columns filled with NA values from the taxonomy table of phyloseq object.
}
\examples{
data(GlobalPatterns)
gp <- tax_glom(GlobalPatterns, taxrank="Family")  # 7 taxonomic ranks
head(tax_table(gp))   # last columns are filled with NAs

# Remove taxonomic levels filled with NAs
gp2 <- phyloseq_rm_na_tax(gp)  # 5 taxonomic ranks
head(tax_table(gp2))

}
\seealso{
\code{\link{tax_table}}
}
