% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_sep_pairwise.R
\name{phyloseq_sep_pairwise}
\alias{phyloseq_sep_pairwise}
\title{Split phyloseq-class object for pairwise comparisons by sample-level variable.}
\usage{
phyloseq_sep_pairwise(phys, group, comparis = "all", drop_zeroes = TRUE)
}
\arguments{
\item{phys}{A phyloseq-class object}

\item{group}{Variable name (contained in \code{\link[phyloseq]{sample_data}})}

\item{comparis}{Character "all" for all possible pairwise comparisons (default), or a matrix with comparison names to use (similar to the output of \code{\link{combn}} with m=2)}

\item{drop_zeroes}{Logical, indicating weather OTUs with zero abundance or samples with zero total abundance should be removed}
}
\value{
List with phyloseq objects (each with only two groups based on selected variable).
}
\description{
Split phyloseq-class object for pairwise comparisons by sample-level variable.
}
\examples{
data(enterotype)

# Split enterotype data for pairwise comparisons by sequencing technology
ent <- phyloseq_sep_pairwise(enterotype, group = "SeqTech", drop_zeroes = TRUE)
ent

# Perform differential abundance testing for each of the pairwise comparisons

}
