% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_sep_samp.R
\name{phyloseq_sep_samp}
\alias{phyloseq_sep_samp}
\title{Separate phyloseq-class object by sample.}
\usage{
phyloseq_sep_samp(physeq, drop_zeroes = T)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{drop_zeroes}{Logical, indicating weather OTUs with zero abundance or samples with zero total abundance should be removed}
}
\value{
List with phyloseq objects.
}
\description{
This function splits a phyloseq object by sample, returning a list of objects whose components each correspond to a separate sample.
}
\examples{
data("esophagus")
phyloseq_sep_samp(esophagus)

}
