% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_sep_variable.R
\name{phyloseq_sep_variable}
\alias{phyloseq_sep_variable}
\title{Split phyloseq-class object by sample-level variable.}
\usage{
phyloseq_sep_variable(physeq, variable, drop_zeroes = T)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{variable}{Variable name (contained in \code{\link[phyloseq]{sample_data}})}

\item{drop_zeroes}{Logical, indicating weather OTUs with zero abundance withing a group of samples should be removed}
}
\value{
List with phyloseq objects.
}
\description{
This function splits a phyloseq object by sample meta-data, returning a list of objects whose components each correspond to a group of samples.
}
\examples{
data(GlobalPatterns)

# Split data by sample type (e.g., Soil, Ocean, etc.)
phyloseq_sep_variable(GlobalPatterns, variable = "SampleType")

# Do not remove OTUs with total zero abundance within each sample type
phyloseq_sep_variable(GlobalPatterns, variable = "SampleType", drop_zeroes = FALSE)

}
