% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_some.R
\name{phyloseq_some}
\alias{phyloseq_some}
\title{Inspect random samples (and taxa) in a phyloseq object}
\usage{
phyloseq_some(x, n = 10, n_otu = 10)
}
\arguments{
\item{x}{A \code{\link[phyloseq]{phyloseq-class}} object}

\item{n}{Number of samples to return}

\item{n_otu}{Optional number of taxa (OTUs/ASVs) to retain; set to \code{NULL}
to keep all taxa}
}
\value{
A phyloseq object pruned to \code{n} samples and up to \code{n_otu} taxa.
}
\description{
Inspect a random subset of samples from a
  \code{\link[phyloseq]{phyloseq-class}} object. Optionally, also
  subsample taxa (OTUs/ASVs) for compact viewing.
}
\examples{
library(phyloseq)
data(enterotype)

phyloseq_some(enterotype)


}
\seealso{
\code{\link[phyloseq]{prune_samples}}, \code{\link[phyloseq]{prune_taxa}},
  \code{\link[phyloseq]{nsamples}}, \code{\link[phyloseq]{ntaxa}}
}
\author{
Inspired by \code{\link[car]{some}} in the car-package by John Fox.
}
