% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_standardize_otu_abundance.R
\name{phyloseq_standardize_otu_abundance}
\alias{phyloseq_standardize_otu_abundance}
\title{Standardize OTU abundance table}
\usage{
phyloseq_standardize_otu_abundance(
  physeq,
  method = "total",
  rhea_depth = NULL,
  rhea_round = TRUE,
  rhea_threshold = NULL,
  ...
)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{method}{Standardization method}

\item{...}{Additional parameters may be passed to vegan \code{\link[vegan]{decostand}} function}
}
\value{
phyloseq object with standardized OTU table.
}
\description{
Standardize phyloseq OTU table with with methods from \code{\link[vegan]{decostand}} from vegan package.
}
\details{
Supported methods:
\itemize{
\item \strong{"total"} - convert data to relative abundances (divide by sample total);
\item \strong{"pa"} - convert OTU abundances to presence/absence scale (0/1);
\item \strong{"log"} - logarithmic transformation as suggested by Anderson et al. (2006): log (x) + 1 for x > 0, zeros are left as zeros, logarithm base = 2. Please note this is not log(x+1);
\item \strong{"hellinger"} - square root of method = "total" (Legendre & Gallagher 2001);
\item \strong{"max"} - divide by sample maximum;
\item \strong{"frequency"} - divide by sample total and multiply by the number of non-zero items, so that the average of non-zero entries is one;
\item \strong{"normalize"} - make sample sum of squares equal to one;
\item \strong{"range"} - standardize values into range 0 ... 1. If all values are constant, they will be transformed to 0;
\item \strong{"rank"} - replace abundance values by their increasing ranks leaving zeros unchanged. Average ranks are used for tied values;
\item \strong{"rrank"} - replace abundance values by relative ranks with maximum 1. Average ranks are used for tied values;
\item \strong{"standardize"} - scale OTU abundances within sample to zero mean and unit variance;
\item \strong{"wisconsin"} - Wisconsin double standardization where species are first standardized by maxima and then sites by site totals;
\item \strong{"chi.square"} - divide by sample sums and square root of OTU sums, and adjust for square root of matrix total (Legendre & Gallagher 2001). When used with the Euclidean distance, the distances should be similar to the Chi-square distance used in correspondence analysis.
}

For the implementation of "total", "max", "frequency", "normalize", "range", "rank", "standardize", "pa", "chi.square", "hellinger", "log" methods see \code{\link[vegan]{decostand}}.
}
\examples{
# Load data
data("esophagus")

# Total-sum scaling (TSS) normalization
phyloseq_standardize_otu_abundance(esophagus, method = "total")
# the same as
transform_sample_counts(esophagus, function(OTU) OTU/sum(OTU) )
identical(
  phyloseq_standardize_otu_abundance(esophagus, method = "total"),
  transform_sample_counts(esophagus, function(OTU) OTU/sum(OTU)) )

# Presence-absence scaling (0/1)
phyloseq_standardize_otu_abundance(esophagus, method = "pa")

# Logarithmic transformation as in Anderson et al., 2006
phyloseq_standardize_otu_abundance(esophagus, method = "log")

# Hellinger standardization
phyloseq_standardize_otu_abundance(esophagus, method = "hellinger")

}
\seealso{
\code{\link[vegan]{decostand}}, \code{\link{phyloseq_transform_css}}, \code{\link{phyloseq_transform_vst_blind}}, \code{\link{phyloseq_transform_rlog_blind}}, \code{\link{physeq_transform_anderson_log}}
}
