% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_to_df.R
\name{phyloseq_to_df}
\alias{phyloseq_to_df}
\title{Convert phyloseq object to data frame (for exporting).}
\usage{
phyloseq_to_df(
  physeq,
  addtax = T,
  addtot = F,
  addmaxrank = F,
  sorting = "abundance"
)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{addtax}{Logical; add taxonomy to the result table (default, TRUE)}

\item{addtot}{Logical; add total OTU abundance to the resulting table (default, FALSE)}

\item{addmaxrank}{Logical; add the OTU classification at the lowest annotated taxonomic rank to the resulting table (default, FALSE)}

\item{sorting}{Reorder OTUs according to their total abundance ("abundance", default) or taxonomy (sorting = "taxonomy"); if reordering is not required use 'sorting = NULL'}
}
\value{
Data frame with OTU taxonomy and abundance
}
\description{
Convert phyloseq object to data frame (for exporting).
}
\examples{
data(GlobalPatterns)
GlobalPatternsDF <- phyloseq_to_df(GlobalPatterns)
str(GlobalPatternsDF)

}
\seealso{
\code{\link[phyloseq]{phyloseq-class}}
}
