% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_transform.R
\name{phyloseq_transform_aldex_clr}
\alias{phyloseq_transform_aldex_clr}
\title{ALDEx2-based centred log-ratio transformation of OTU table}
\usage{
phyloseq_transform_aldex_clr(physeq, variable = NULL, iter = 1)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{variable}{Variable name (from \code{\link[phyloseq]{sample_data}}) containing a descriptor for the samples, allowing them to be grouped and compared}

\item{iter}{Number of Monte Carlo samples to use}
}
\value{
Phyloseq object with CLR-transformed OTU counts (or a list of rarefied phyloseq-objects if iter > 1).
}
\description{
ALDEx2-based centred log-ratio transformation of OTU table
}
\details{
ALDEx2 generates Monte Carlo samples of the Dirichlet distribution for each sample and
converts each instance using the centred log-ratio transformation.
}
\references{
Fernandes DA, Reid J, Macklaim MJ, McMurrough TA, Edgell DR, Gloor BG (2014) Unifying the analysis of high-throughput sequencing datasets: characterizing RNA-seq, 16S rRNA gene sequencing and selective growth experiments by compositional data analysis // Microbiome 2:15. DOI: 10.1186/2049-2618-2-15
}
\seealso{
\code{\link[ALDEx2]{aldex.clr}}
}
