% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_transform.R
\name{phyloseq_transform_css}
\alias{phyloseq_transform_css}
\title{Cumulative sum scaling (CSS) normalization of OTU abundance table.}
\usage{
phyloseq_transform_css(physeq, norm = TRUE, log = TRUE, ...)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{norm}{Logical, return normalized counts}

\item{log}{Logical, apply a logarithmic transform (log2) to the normalized count data}

\item{...}{Additional arguments will be passed to \code{\link{MRcounts}}}
}
\value{
Phyloseq object with transformed counts in OTU table.
}
\details{
Median scaling factor across samples will be used as default.
}
\examples{

}
\references{
Paulson et al. Nature Methods 10, 1200–1202 (2013) doi:10.1038/nmeth.2658. http://www.nature.com/nmeth/journal/v10/n12/full/nmeth.2658.html
}
\seealso{
\code{\link{phyloseq_transform_vst_blind}}, \code{\link{phyloseq_transform_rlog_blind}}, \code{\link{rlog}}, \code{\link{varianceStabilizingTransformation}}
}
