% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_transform.R
\name{phyloseq_transform_rlog_blind}
\alias{phyloseq_transform_rlog_blind}
\title{Regularized-log (rlog) transformation of OTU abundance table.}
\usage{
phyloseq_transform_rlog_blind(physeq, dropneg = F, dropmissing = T, ...)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{dropneg}{Logical, replace negative transformed values with 0}

\item{dropmissing}{Logical, remove missing data}

\item{...}{Additional arguments will be passed to \code{\link[DESeq2]{rlogTransformation}}}
}
\value{
Phyloseq object with transformed counts in OTU table.
}
\description{
Regularized-log (rlog) transformation of OTU abundance table.
}
\details{
rlog transformation (this function) is preferable to the vst (\code{\link{phyloseq_transform_vst_blind}}) if the size factors vary widely.
}
\seealso{
\code{\link{phyloseq_transform_vst_blind}}, \code{\link{phyloseq_transform_css}}, \code{\link[DESeq2]{rlog}}, \code{\link[DESeq2]{varianceStabilizingTransformation}}
}
