% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_transform.R
\name{phyloseq_transform_vst_blind}
\alias{phyloseq_transform_vst_blind}
\title{Variance stabilizing transformation (VST) of OTU abundance table.}
\usage{
phyloseq_transform_vst_blind(physeq, dropneg = F, dropmissing = T, ...)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{dropneg}{Logical, replace negative transformed values with 0}

\item{dropmissing}{Logical, remove missing data}

\item{...}{Not yet implemented}
}
\value{
Phyloseq object with transformed counts in OTU table.
}
\description{
Variance stabilizing transformation (VST) of OTU abundance table.
}
\details{
For downstream analysis it could be better to use sample covariate information (blind = FALSE in \code{\link[DESeq2]{varianceStabilizingTransformation}}).
}
\seealso{
\code{\link{phyloseq_transform_rlog_blind}}, \code{\link{phyloseq_transform_css}}, \code{\link[DESeq2]{varianceStabilizingTransformation}}, \code{\link[DESeq2]{rlog}}
}
