% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_transform.R
\name{physeq_transform_anderson_log}
\alias{physeq_transform_anderson_log}
\title{Log-transformation of OTU abundance table.}
\usage{
physeq_transform_anderson_log(physeq, ...)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{...}{Additional arguments (e.g., "logbase" for the logarithm base) will be passed to \code{\link{decostand}}}
}
\value{
Phyloseq object with log-transformed counts in OTU table.
}
\description{
Log-transformation of OTU abundance table.
}
\details{
Logarithmic transformation as suggested by Anderson et al. (2006) will be applied to the OTU table in phyloseq-object. First, non-integer data will be divided by smallest positive value. Second, log(x) + 1 for x > 0.
Default value of the logarithm base ("logbase" parameter) is 2.
This function is a wrapper to \code{\link{decostand}} in vegan.
}
\examples{
data(enterotype)
physeq_transform_anderson_log(enterotype)

}
\references{
Anderson MJ, Ellingsen KE and McArdle BH. (2006) Multivariate dispersion as a measure of beta diversity. Ecology Letters 9, P. 683–693.
}
\seealso{
\code{\link{decostand}}
}
