% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_breakaway.R
\name{prepare_breakaway}
\alias{prepare_breakaway}
\title{Prepare frequency counts for breakaway package}
\usage{
prepare_breakaway(OTUs, remove_singletons = FALSE)
}
\arguments{
\item{OTUs}{Data frame with OTU abundances (rows = species, cols = samples) or phyloseq/otu_table object}

\item{remove_singletons}{Logical; if TRUE, singleton counts will be removed}
}
\value{
List of the sample frequency count tables for each sample.
}
\description{
This function prepares input data for \code{\link[breakaway]{breakaway}} package (Willis, Bunge, 2015).
}
\details{
In case of sequencing errors singletons can be excluded from the data ('remove_singletons = TRUE')
and species richness estimation can be performed with \code{\link[breakaway]{breakaway_nof1}} function.
}
\examples{
data("esophagus")
eso <- prepare_breakaway(esophagus)
eso_nof1 <- prepare_breakaway(esophagus, remove_singletons = TRUE)

# Estimate species richness (observed + unobserved) for the second sample in esophagus data
library(breakaway)
breakaway(eso[[2]], plot = T)
breakaway_nof1(eso_nof1[[2]], plot = T)

}
\references{
Willis A., Bunge J. (2015) Estimating diversity via frequency ratios // Biometrics 71(4), P. 1042–1049. DOI: 10.1111/biom.12332
}
