% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_m8.R
\name{read_m8}
\alias{read_m8}
\title{Read BLAST output}
\usage{
read_m8(x, blast_colz = NULL, package = "data.table")
}
\arguments{
\item{x}{Input file name}

\item{blast_colz}{Vector with column names (optional)}

\item{package}{Package to use for data preparation (default, "data.table")}
}
\value{
data.table / data.frame / tibble (depending on the `package` parameter)
}
\description{
This function loads BLAST results saved in a tabular output format 6
(`-outfmt 6`; 12 columns).
}
\examples{
# read_m8("blast_results.m8")

}
