% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SES.R
\name{SES}
\alias{SES}
\title{Estimate standardized effect size (SES) and p-value}
\usage{
SES(obs, sim, alternative = "two-sided")
}
\arguments{
\item{obs}{Single value of the observed statistics}

\item{sim}{Numeric vector of simulated values (null distribution)}

\item{alternative}{Character string specifying the alternative hypothesis; "greater", "less", or "two-sided" (default)}
}
\value{
Data frame with the effect size estimate (SES) and p-value (Pval).
}
\description{
Estimate standardized effect size (SES) and p-value
}
\details{
Standardized effect size (SES) represents Z-score, and is calculated as
(observed - mean_simulated) / standard_deviation_simulated.
SES values more extreme than +/-1.96 belong to the outer 5% of the null distribution
and thus are significant at alpha=0.05 with two-sided hypothesis test.
}
\examples{
SES(2, 1:10)

}
