% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some.R
\name{some}
\alias{some}
\title{Check random lines in the data (including phyloseq objects)}
\usage{
some(x, n = 10, n_otu = 10)
}
\arguments{
\item{x}{Coluld be a \code{\link{data.frame}}, \code{\link{matrix}}, or \code{\link[phyloseq]{phyloseq-class}}}

\item{n}{Number of records (or samples in phyloseq) that will be returned}

\item{n_otu}{Number of OTUs to show (if the input is phyloseq-object)}
}
\value{
Part of the data.frame or matrix with n lines or a subset of phyloseq-object.
}
\description{
Returns several random lines from the data.frame, matrix or phyloseq.
Usage is quite similar to the \code{\link{head}} or \code{\link{tail}} functions.
}
\examples{
# Load phyloseq data
data(enterotype)
some(enterotype)

# If input is data.frame or matrix
x <- matrix(rnorm(500), ncol=5)
some(x)

}
\author{
Adapted from \code{\link[car]{some}} in the car-package by John Fox.
}
