% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_to_qiime.R
\name{taxonomy_to_qiime}
\alias{taxonomy_to_qiime}
\title{Prepare taxonomy in QIIME-style}
\usage{
taxonomy_to_qiime(x, dropNA = TRUE, add_OTUID = TRUE, custom_tax_ranks = NULL)
}
\arguments{
\item{x}{data.frame or phyloseq with \code{\link[phyloseq]{tax_table}} slot}

\item{dropNA}{Logical; if TRUE, missing tax ranks will be removed}

\item{add_OTUID}{Logical; if TRUE, OTU name will be added to the first column of the resulting table}

\item{custom_tax_ranks}{Data frame with two columns, taxonomic ranks (with the same names as columns in x) and the corresponding single letter abbreviations}
}
\value{
character vector (if 'add_OTUID = FALSE') or data.frame with taxonomy annotations.
}
\description{
This function merges multiple columns with taxonomic ranks into a single string formatted in QIIME style.
}
\details{
If 'add_OTUID = T' and x is data.frame, the function will take OTU names from rownames of a data frame.
}
\examples{
# Load data
data("GlobalPatterns")

# Subset data
GP <- metagMisc::phyloseq_filter_top_taxa(GlobalPatterns, n = 20)

# Prepare taxonomy in QIIME-style (result = data frame with 2 columns)
qtax <- taxonomy_to_qiime(GP)
head(qtax)

# Vector with OTU taxonomy
qvec <- taxonomy_to_qiime(GP, add_OTUID = FALSE)
head(qvec)

}
