% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_to_sumaclust.R
\name{uc_to_sumaclust}
\alias{uc_to_sumaclust}
\title{Convert UC file (clustering output from USEARCH and VSEARCH) to Sumaclust OTU observation map}
\usage{
uc_to_sumaclust(x, output_file = NULL)
}
\arguments{
\item{x}{File name in UC-format (typically with .uc extension)}

\item{output_file}{Optional, character string with the name of the output file}
}
\value{
This function invisibly returns a list, where each slot corresponds to a single OTU.
Each slot is a charater vector that contains sequence names that belongs to this OTU.
The first element of a vector is a name of the representative sequence of this OTU (OTU centoid)
which occurs twice in this vector (i.e., there are two identical sequence IDs in a vector
for OTUs that are represented by only one sequence).
}
\description{
Convert UC file (clustering output from USEARCH and VSEARCH) to Sumaclust OTU observation map
}
\examples{
## OTU clustering with Sumaclust
## OTU observation map is written to 'sumaclust_map.txt'
# sumaclust -g -f -t 0.97 -O sumaclust_map.txt input_seqs.fasta

## OTU clustering with VSEARCH
## Results are in USEARCH cluster format (UC) in 'vsearch_clusters.uc' file
# vsearch --cluster_fast input_seqs.fasta --id 0.97 --uc vsearch_clusters.uc

## Convert VSEARCH clustering results (in UC-format) to Sumaclust OTU observation map
# uc_to_sumaclust("vsearch_clusters.uc", output_file = "vsearch_to_sumaclust.txt")

}
\references{
https://git.metabarcoding.org/obitools/sumaclust/wikis/home
}
\seealso{
\code{\link{parse_uc}}
}
