% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_taxa_names.R
\name{abbreviate_taxa_names}
\alias{abbreviate_taxa_names}
\title{Abbreviate taxa names}
\usage{
abbreviate_taxa_names(names, nlet = 3, totl = 7, sep = "_", seconditem = F)
}
\arguments{
\item{names}{Character vector of species names}

\item{nlet}{Number of letters to take from genus-species parts}

\item{totl}{Number of letters in the final abbreviation}

\item{sep}{Which separator to use for genus-species case (default is underscore)}

\item{seconditem}{Logical, take the second item of the original name for abbreviation}
}
\value{
Character vector of abbreviated taxa names
}
\description{
This function is based on the deprecated vegan::make.cepnames function 
(git 1b816c1; Aug 2, 2011) and provides similar functionality for abbreviating taxa names.
}
\examples{
x <- c("Laccaria laccata", "Meliniomyces bicolor",
  "Inocybe cincinnata", "Inocybe", "Tylospora asterophora",
  "Cadophora finlandica", "Saccharomycetales")

abbreviate_taxa_names(x, nlet = 3, totl = 7, sep = "_")
abbreviate_taxa_names(x, nlet = 4, totl = 8, sep = "")

}
\author{
Author of the original function is Jari Oksanen (vegan::make.cepnames)
}
