% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to data.frame}
\usage{
add_metadata(x, metad, xid = NULL, mid = NULL, drop_mid = T)
}
\arguments{
\item{x}{Data frame with data of interest}

\item{metad}{Data frame with metadata that should be added; or phyloseq object with sample_data slot}

\item{xid}{Character, name of the column with row identifiers in the main data (x)}

\item{mid}{Character, name of the column with row identifiers in the metadata (metad)}

\item{drop_mid}{Logical; if TRUE (default), metadata identifiers will be removed from results}
}
\value{
Data frame
}
\description{
This function combines two tables based on the row identifiers.
}
\examples{
# Load data
data(GlobalPatterns)

# As an example, estimate richness
div <- estimate_richness(GlobalPatterns, measures = c("Observed", "Shannon"))
div <- dfRowName(div, name = "SampleID")

# Take metadata from phyloseq object
div1 <- add_metadata(div, metad = GlobalPatterns, xid = "SampleID", mid = "X.SampleID")
head(div1)

# Or metadata from data frame
metad <- as(sample_data(GlobalPatterns), "data.frame")
div2 <- add_metadata(div, metad = metad, xid = "SampleID", mid = "X.SampleID")
head(div2)

}
