% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis_pairwise.R
\name{adonis_pairwise}
\alias{adonis_pairwise}
\title{Pairwise comparisons for permutational multivariate analysis of variance using distance matrices}
\usage{
adonis_pairwise(x, dd, group.var, add_permdisp = TRUE, permut = 999,
  p.adj = T, adj.meth = "fdr", all_results = T, comparison_sep = ".",
  permdisp_type = "median", ...)
}
\arguments{
\item{x}{Sample meta-data (data frame for the independent variables)}

\item{dd}{Dissimilarity matrix between samples}

\item{group.var}{Name of the independent variable to test (RHS in adonis formula)}

\item{add_permdisp}{Logical; if TRUE (default), results of tests for homogeneity of multivariate dispersions will be added to output (see \code{\link[vegan]{betadisper}})}

\item{permut}{Number of permutations required}

\item{p.adj}{Logical, adjust P-values for multiple comparisons}

\item{adj.meth}{Correction method from \code{\link{p.adjust}} ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr")}

\item{all_results}{Logical, return results of adonis and data subsets for each pairwise comparison}

\item{comparison_sep}{Character string to separate the levels of independent variable the in the pairwise comparison names (default, ".")}

\item{permdisp_type}{Use the spatial median (default) or the group centroid for the analysis for homogeneity of multivariate dispersions (see \code{\link[vegan]{betadisper}})}

\item{...}{Additional arguments will be passed to \code{\link[vegan]{adonis}}}
}
\value{
List with adinonis and betadisper results.
}
\examples{
library(vegan)
data(dune)
data(dune.env)

# Compare all Management levels
adonis(dune ~ Management, data = dune.env)

# Pairwise comparisons between Management levels
tst <- adonis_pairwise(x = dune.env, dd = vegdist(dune), group.var = "Management")
tst$Adonis.tab
tst$Betadisper.tab

}
\seealso{
\code{\link[vegan]{adonis}}, \code{\link[vegan]{betadisper}}
}
