% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_amptk_taxonomy_table.R
\name{parse_taxonomy_amptk}
\alias{parse_taxonomy_amptk}
\title{Parse taxonomy string from AMPtk (for single OTU)}
\usage{
parse_taxonomy_amptk(x)
}
\arguments{
\item{x}{Character vector of length 1 with taxonomic annotation of single OTU/species.}
}
\value{
Named character vector with taxonomic ranks of OTU or species.
}
\details{
This function will split
}
\examples{
parse_taxonomy_amptk("k:Fungi,p:Zygomycota,o:Mortierellales,f:Mortierellaceae,g:Mortierella,s:Mortierella parvispora")
parse_taxonomy_amptk("k:Fungi,p:Ascomycota,g:Chalara")

}
\references{
AMPtk: Amplicon ToolKit for NGS data (formally UFITS): https://github.com/nextgenusfs/amptk
}
\seealso{
\code{\link{parse_taxonomy_qiime}}
}
