% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_amptk_taxonomy_table.R
\name{parse_taxonomy_amptk_batch}
\alias{parse_taxonomy_amptk_batch}
\title{Parse multiple taxonomy strings from AMPtk.}
\usage{
parse_taxonomy_amptk_batch(x, withID = TRUE, multithread = FALSE)
}
\arguments{
\item{x}{Vector of character strings from AMPtk taxonomy}

\item{withID}{Logical; set to TRUE (default) if charter strings contain sequence IDs}

\item{multithread}{Logical; run the function in parallel}
}
\value{
Data frame with OTUs or species as rows and their taxonomic ranks as columns.
}
\description{
Parse multiple taxonomy strings from AMPtk.
}
\details{
This function splits AMPtk (ex-UFITS) result of taxonomic annotation by taxonomic rank and combines it into one table.
The first part of taxonomy string is assumed to be an ID of the best hit (e.g., EF040844) or a taxonomic classificator (UTAX or SINTAX) followed by a semicolon.
}
\examples{
## Taxonomy with sequence ID
tax <- c(
"UDB011192;k:Fungi,p:Basidiomycota,c:Agaricomycetes,o:Russulales,f:Russulaceae,g:Russula,s:Russula chloroides",
"SINTAX;k:Fungi,p:Chytridiomycota,c:Chytridiomycetes,o:Rhizophydiales",
"EF040844;k:Fungi,p:Basidiomycota,c:Agaricomycetes,o:Atheliales,f:Atheliaceae,g:Piloderma",
"EU240039;k:Fungi,p:Zygomycota,o:Mortierellales,f:Mortierellaceae,g:Mortierella",
"UTAX;k:Fungi,p:Basidiomycota,c:Agaricomycetes,o:Atheliales,f:Atheliaceae,g:Amphinema,s:Amphinema byssoides",
"SINTAX;k:Fungi,p:Chytridiomycota,c:Chytridiomycetes,o:Spizellomycetales"
)

parse_taxonomy_amptk_batch(tax)

## Taxonomy without sequence ID
tax2 <- c(
"k:Fungi,p:Basidiomycota,c:Agaricomycetes,o:Russulales,f:Russulaceae,g:Russula,s:Russula chloroides",
"k:Fungi,p:Chytridiomycota,c:Chytridiomycetes,o:Rhizophydiales",
"k:Fungi,p:Basidiomycota,c:Agaricomycetes,o:Atheliales,f:Atheliaceae,g:Piloderma",
"k:Fungi,p:Zygomycota,o:Mortierellales,f:Mortierellaceae,g:Mortierella",
"k:Fungi,p:Basidiomycota,c:Agaricomycetes,o:Atheliales,f:Atheliaceae,g:Amphinema,s:Amphinema byssoides",
"k:Fungi,p:Chytridiomycota,c:Chytridiomycetes,o:Spizellomycetales"
)

parse_taxonomy_amptk_batch(tax2, withID = FALSE)

}
\references{
AMPtk: Amplicon ToolKit for NGS data (formally UFITS): https://github.com/nextgenusfs/amptk
}
\seealso{
\code{\link{parse_taxonomy_amptk}} \code{\link{parse_taxonomy_qiime}}
}
