% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_combine_samples.R
\name{phyloseq_combine_samples}
\alias{phyloseq_combine_samples}
\title{Combine samples by summing OTU abundance}
\usage{
phyloseq_combine_samples(physeq, variable = NULL)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{variable}{Character string; name of the sample metadata variable to use for grouping. 
If NULL (default), all samples are combined into a single sample.}
}
\value{
A phyloseq object with combined samples. Original sample metadata is removed
  when combining all samples, but taxonomy table and phylogenetic tree (if present) are preserved.
}
\description{
Sums OTU abundances of replicates or sample groups into single samples.
This function can either combine all samples into one single sample or 
combine samples within groups defined by a sample metadata variable.
}
\examples{
\dontrun{
data("GlobalPatterns")

# Combine all samples into a single sample
gp_total <- phyloseq_combine_samples(GlobalPatterns)

# Combine samples by SampleType
gp_by_type <- phyloseq_combine_samples(GlobalPatterns, variable = "SampleType")
}

}
\seealso{
\code{\link{phyloseq_merge_samples}}, \code{\link{phyloseq_sep_variable}}
}
