% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_compare.R
\name{phyloseq_compare}
\alias{phyloseq_compare}
\title{Compare two phyloseq objects.}
\usage{
phyloseq_compare(phy1, phy2, cols = c("Before", "After"))
}
\arguments{
\item{phy1}{Phyloseq object}

\item{phy2}{Phyloseq object}

\item{cols}{Character vector with column names for phy1 & phy2 in the resulting table.}
}
\value{
Data frame with number and percentage of OTUs.
}
\examples{
data(esophagus)

# Remove taxa that have less than 5 reads
eso_trim <- prune_taxa(taxa_sums(esophagus) >= 5, esophagus)

phyloseq_compare(esophagus, eso_trim, cols = c("Esophagus", "Trimmed esophagus"))

}
