% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_coverage.R
\name{phyloseq_coverage}
\alias{phyloseq_coverage}
\title{Estimate the observed abundance-based sample coverage for phyloseq object}
\usage{
phyloseq_coverage(physeq, correct_singletons = FALSE, add_attr = T)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{correct_singletons}{Logical; if TRUE, singleton counts will be corrected with modified Good–Turing frequency formula (Chiu, Chao 2016)}

\item{add_attr}{Logical; if TRUE, additional attributes (list of species abundances and singleton correction flag) will be added to the results}
}
\value{
Data frame with coverage estimates for each sample
}
\description{
phyloseq_coverage estimates the sample completeness for the individual-based
abundance data (number of sequencing reads) stored in 'phyloseq'-class objects.
}
\details{
Coverage represents a measure of sample completeness and is defined as the proportion of
the total number of individuals in a community that belong to the species represented in the sample.
Coverage complement (1 - Coverage) gives the proportion of the community belonging to unsampled
species or the "coverage deficit" (Chao, Jost, 2012).
Estimation of coverage is based on the number of singletons and doubletons in the sample.
}
\examples{
data("esophagus")
phyloseq_coverage(esophagus)

}
\references{
Chao A, Jost L. (2012) Coverage-based rarefaction and extrapolation: standardizing samples by completeness rather than size // Ecology 93(12): 2533–2547. DOI: 10.1890/11-1952.1
}
