% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_filter.R
\name{phyloseq_filter_prevalence}
\alias{phyloseq_filter_prevalence}
\title{Filter low-prevalence OTUs.}
\usage{
phyloseq_filter_prevalence(physeq, prev.trh = 0.05, abund.trh = NULL)
}
\arguments{
\item{physeq}{Phyloseq object}

\item{prev.trh}{Prevalence threshold (default, 0.05 = 5\% of samples)}

\item{abund.trh}{Abundance threshold (default, NULL)}
}
\value{
Phyloseq object with a subset of taxa.
}
\description{
This function will remove taxa (OTUs) with low prevalence, where prevalence is the fraction of total samples in which an OTU is observed.
}
\details{
Abundance threshold defines if the OTU should be preserved if its abundance is larger than threshold (e.g., >= 50 reads).
}
\examples{
data(GlobalPatterns)
GlobalPatterns  # 19216 taxa

# OTUs that are found in at least 5\% of samples
phyloseq_filter_prevalence(GlobalPatterns, prev.trh = 0.05, abund.trh = NULL)  # 15389 taxa

# The same, but if total OTU abundance is >= 10 reads it'll be preserved too
phyloseq_filter_prevalence(GlobalPatterns, prev.trh = 0.05, abund.trh = 10)    # 15611 taxa

}
