% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_phylo_ses.R
\name{phyloseq_phylo_ses}
\alias{phyloseq_phylo_ses}
\title{Estimate standardized effect sizes (SES) of phylogenetic diversity metrics (PD, MPD, MNTD, VPD) using randomization-based approach.}
\usage{
phyloseq_phylo_ses(physeq, measures = c("PD", "MPD", "MNTD", "VPD"),
  null_model = "taxa.labels", package = "picante",
  abundance_weighted = FALSE, nsim = 1000, swapiter = 1000,
  verbose = TRUE, ...)
}
\arguments{
\item{physeq}{A phyloseq-class object (phylogenetic tree is required)}

\item{measures}{Character vector with diversity indices names ("PD", "MPD", "MNTD", "VPD")}

\item{null_model}{Character string indicating which null model to use (for the supported list of models see \code{\link[picante]{ses.pd}})}

\item{package}{Which package to use for diversity estimation - "picante" (default) or "PhyloMeasures" (could be faster)}

\item{abundance_weighted}{Logical; should diversity metrics be weighted by species abundance? (default = FALSE)}

\item{nsim}{Number of randomizations for null-distribution generation}

\item{swapiter}{Number of iterations for independentswap or trialswap algorithms}

\item{verbose}{Logical; if TRUE, progress messages from the function will be printed}

\item{...}{Additional arguments currently are not implemented}
}
\value{
Data frame.
}
\description{
This function utilizes randomization-based approach to computet SES values
(in contrast to the exact standardisation approach used in \code{\link{phyloseq_phylo_div}} with 'standardize' option),
which could be more computation intensive. Note that despite more freedom in defining the null-model with randomization-based approach
it is debatable that this method can provide a reasonable approximation by selecting only a limited number of possible combinations
(especially for large phylogenetic trees).

Currently only non-abundance-weighted estimates are implemented.
}
\details{
Available metrics include:
\itemize{
  \item PD - Faith's phylogenetic diversity (total length of all the phylogenetic branches for a given set of taxa);
  \item MPD - Mean pairwise distance between the tips from a phylogenetic tree;
  \item MNTD - Mean nearest taxon distance;
  \item VPD - Variance of the pairwise distances.
}
Values of SES above zero indicate that the species pool of a habitat is more
diverse than the regional species pool.
}
\examples{
# Load data
data(esophagus)

# Estimate phylogenetic diversity index (PDI)
phyloseq_phylo_ses(esophagus, measures = "PD", null_model = "taxa.labels", nsim = 100, verbose = F)

# Estimate phylogenetic diversity, mean pairwise distance,
# mean nearest taxon distance and variance of phylogenetic distances;
# use independent swap algorithm to generate the null distribution
phyloseq_phylo_ses(esophagus,
                   measures = c("PD", "MPD", "MNTD", "VPD"),
                   null_model = "independentswap",
                   nsim = 200, swapiter = 200)  # NB. increase the number of iterations!

}
\references{
Webb C., Ackerly D., McPeek M., Donoghue M. (2002) Phylogenies and community ecology. Annual Review of Ecology and Systematics, 33, 475-505.
}
\seealso{
\code{\link{phyloseq_phylo_div}}, \code{\link{phyloseq_randomize}},
\code{\link[picante]{randomizeMatrix}}, \code{\link[picante]{ses.pd}},
\code{\link[picante]{ses.mpd}}, \code{\link[picante]{ses.mntd}},
\code{\link[PhyloMeasures]{pd.query}}, \code{\link[PhyloMeasures]{mpd.query}},
\code{\link[PhyloMeasures]{mntd.query}}
}
