% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_standardize_otu_abundance.R
\name{phyloseq_standardize_otu_abundance}
\alias{phyloseq_standardize_otu_abundance}
\title{Standardize OTU abundance table}
\usage{
phyloseq_standardize_otu_abundance(physeq, method = "total", ...)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{method}{Standardization method (see \code{\link[vegan]{decostand}} for available options)}

\item{...}{Additional parameters may be passed to vegan \code{\link[vegan]{decostand}} function}
}
\value{
phyloseq object with standardized OTU table.
}
\description{
Standardize phyloseq OTU table with with methods from \code{\link[vegan]{decostand}} from vegan package.
}
\examples{
# Load data
data("esophagus")

# Total-sum scaling (TSS) normalization
phyloseq_standardize_otu_abundance(esophagus, method = "total")
# the same as
transform_sample_counts(esophagus, function(OTU) OTU/sum(OTU) )
identical(
  phyloseq_standardize_otu_abundance(esophagus, method = "total"),
  transform_sample_counts(esophagus, function(OTU) OTU/sum(OTU)) )

# Presence-absence scaling (0/1)
phyloseq_standardize_otu_abundance(esophagus, method = "pa")

# Logarithmic transformation as in Anderson et al., 2006
phyloseq_standardize_otu_abundance(esophagus, method = "log")

# Hellinger standardization
phyloseq_standardize_otu_abundance(esophagus, method = "hellinger")

}
\seealso{
\code{\link[vegan]{decostand}}, \code{\link{phyloseq_transform_css}}, \code{\link{phyloseq_transform_vst_blind}}, \code{\link{phyloseq_transform_rlog_blind}}, \code{\link{physeq_transform_anderson_log}}
}
