% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_summary.R, R/phyloseq_summary3.R
\name{phyloseq_summary}
\alias{phyloseq_summary}
\alias{phyloseq_summary}
\title{Summary for phyloseq objects.}
\usage{
phyloseq_summary(physeq, ..., cols = NULL, more_stats = FALSE,
  long = FALSE)

phyloseq_summary(physeq, ..., cols = NULL, more_stats = FALSE,
  long = FALSE)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{...}{Optionally more phyloseq objects}

\item{cols}{Character vector with column names for phy1 & phy2 in the resulting table}

\item{more_stats}{Logical; if TRUE, some additional OTU abundance statistics will be calculated}

\item{long}{Logical; if TRUE, results will be returned in a long format}

\item{physeq}{A phyloseq-class object}

\item{...}{Optionally more phyloseq objects}

\item{cols}{Character vector with column names for phy1 & phy2 in the resulting table}

\item{more_stats}{Logical; if TRUE, some additional OTU abundance statistics will be calculated}

\item{long}{Logical; if TRUE, results will be returned in a long format}
}
\value{
Data frame with number and percentage of OTUs.

Data frame with number and percentage of OTUs.
}
\description{
This function could be used for comparison of multiple phyloseq objects (e.g., after trimming).

This function could be used for comparison of multiple phyloseq objects (e.g., after trimming).
}
\details{
Optionally (if more_stats = TRUE), some additional OTU abundance statistics may be estimated (min, max, median, Q1, Q3, and CQV of OTU abundance).
Coefficient of quartile variation (a.k.a., Quartile coefficient of dispersion) is preferred to the 'classical' coefficient of variation for the analysis of samples from nonnormal distributions.
It is estimated by the following formula: (Q3-Q1)/(Q3+Q1), where Q1 is the population 25th percentile and Q3 is the population 75th percentile.

Optionally (if more_stats = TRUE), some additional OTU abundance statistics may be estimated (min, max, median, Q1, Q3, and CQV of OTU abundance).
Coefficient of quartile variation (a.k.a., Quartile coefficient of dispersion) is preferred to the 'classical' coefficient of variation for the analysis of samples from nonnormal distributions.
It is estimated by the following formula: (Q3-Q1)/(Q3+Q1), where Q1 is the population 25th percentile and Q3 is the population 75th percentile.
}
\examples{
data(esophagus)

# Remove taxa that have less than 5 reads
eso_trim <- prune_taxa(taxa_sums(esophagus) >= 5, esophagus)

phyloseq_summary(esophagus, eso_trim, cols = c("Esophagus", "Trimmed esophagus"))
phyloseq_summary(esophagus, eso_trim, cols = c("Esophagus", "Trimmed esophagus"), more_stats = T)

data(esophagus)

# Remove taxa that have less than 5 reads
eso_trim <- prune_taxa(taxa_sums(esophagus) >= 5, esophagus)

phyloseq_summary(esophagus, eso_trim, cols = c("Esophagus", "Trimmed esophagus"))
phyloseq_summary(esophagus, eso_trim, cols = c("Esophagus", "Trimmed esophagus"), more_stats = T)

}
