% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_to_df.R
\name{phyloseq_to_df}
\alias{phyloseq_to_df}
\title{Convert phyloseq object to data frame (for exporting).}
\usage{
phyloseq_to_df(physeq, addtax = T, addtot = F, abund_sort = T)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{addtax}{Add taxonomy to the result table}

\item{addtot}{Add total OTU abundance to the result table}

\item{abund_sort}{Sort OTUs by total abundance}
}
\value{
Data frame with OTU taxonomy and abundance
}
\examples{
data(GlobalPatterns)
GlobalPatternsDF <- phyloseq_to_df(GlobalPatterns)
str(GlobalPatternsDF)

}
\seealso{
\code{\link{phyloseq-class}}
}
