% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Compute prevalence and abundance summary of each species/OTU.}
\usage{
prevalence(physeq, add_tax = TRUE)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{add_tax}{Logical, add taxonomy to the results}
}
\value{
Data frame
}
\examples{
data(esophagus)
prevalence(esophagus)

data(GlobalPatterns)
head( prevalence(GlobalPatterns, add_tax = T) )

}
