% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some.R
\name{some}
\alias{some}
\title{Check random lines in the data}
\usage{
some(x, n = 10)
}
\arguments{
\item{x}{A data.frame or matrix}

\item{n}{Number of records to show}
}
\value{
Part of the data.frame or matrix with n lines.
}
\description{
Returns several random lines from the data.frame or matrix.
Usage is quite similar to the \code{head} or \code{tail} functions.
}
\examples{
x <- matrix(rnorm(500), ncol=5)
some(x)

}
\author{
Adapted from \code{some} in the car-package by John Fox.
}
